+++ src/auth/auth.module.ts
@@ -10,6 +10,11 @@ import { InviteCodeEntity } from './entities/invite-code.entity';
 import { PasswordResetTokenEntity } from './entities/password-reset-token.entity';
 import { JwtStrategy } from './jwt.strategy';
 import { AuthGuard } from './auth.guard';
+import { AuthConfigEntity } from './auth-config.entity';
+import { AuthConfigService } from './auth-config.service';
+import { AdminGuard } from './admin.guard';
+import { UsersController } from './users.controller';
+import { AuthSettingsController } from './auth-settings.controller';
 
 @Module({
   imports: [
@@ -19,10 +24,18 @@ import { AuthGuard } from './auth.guard';
       UserEntity,
       InviteCodeEntity,
       PasswordResetTokenEntity,
+      AuthConfigEntity,
     ]),
   ],
-  controllers: [AuthController],
-  providers: [AuthService, MailerService, JwtStrategy, AuthGuard],
-  exports: [AuthGuard],
+  controllers: [AuthController, UsersController, AuthSettingsController],
+  providers: [
+    AuthService,
+    MailerService,
+    JwtStrategy,
+    AuthGuard,
+    AuthConfigService,
+    AdminGuard,
+  ],
+  exports: [AuthGuard, AuthConfigService],
 })
 export class AuthModule {}