+++ src/auth/entities/user.entity.ts
@@ -14,6 +14,9 @@ export class UserEntity {
   @Column()
   passwordHash!: string;
 
+  @Column({ default: false })
+  isAdmin!: boolean;
+
   @Column()
   createdAt!: string;