+++ src/settings/settings.module.ts
@@ -2,9 +2,10 @@ import { Module, forwardRef } from '@nestjs/common';
 import { SettingsController } from './settings.controller';
 import { SettingsService } from './settings.service';
 import { BookdarrModule } from '../bookdarr/bookdarr.module';
+import { AuthModule } from '../auth/auth.module';
 
 @Module({
-  imports: [forwardRef(() => BookdarrModule)],
+  imports: [forwardRef(() => BookdarrModule), forwardRef(() => AuthModule)],
   controllers: [SettingsController],
   providers: [SettingsService],
   exports: [SettingsService],